// packages/productive-api/src/formatters/types.ts
type JsonApiMeta = ProductiveApiMeta;
// ...
export interface FormattedPagination {
  page: number;
  total_pages: number;
  total_count: number;
}